requirements.txt
fastapi==0.115.0
SQLAlchemy==2.0.35
pydantic==2.8.2
bcrypt==4.2.0
passlib==1.7.4
PyJWT==2.9.0
uvicorn==0.31.1
python-jose==3.3.0
python-dotenv==1.0.1
cryptography==43.0.1
pymysql==1.1.1
aiomysql==0.2.0
aiosqlite==0.20.0